Enable Ec2 datasource for internal artifacts
For our internally used VM artifacts, we want to enable the Ec2
datasource, such that cloud-init will populate the default user's SSH
authorized_keys file when running in AWS EC2. This changes enables this
functionality.

Additionally, it removes the NoCloud datasource, as we don't currently
use that functionality with our internal images, and don't want to
expose that in our external images.