Add new "ansiblecheck" make target
This change adds another make target which leverages the "ansible-lint"
utility to provide some static analysis of our Ansible roles.

The roles don't currently adhere to the default set of rules of this
utility, so many rules have to be added to the "skip list" in order for
this check to run clean.