Clone master branch of product repositories
The Ansible roles that clone product repositories create a local branch
named "projects/dx4linux". Our Jenkins jobs fetch a specific refspec in
order to reduce the amount of time it takes to clone the repository.
When a local branch exists with the same name as this refspec, Jenkins
finds multiple candidate revisions when determining the revision to
build. It then chooses the head of local branch as the revision to
build, and it ignores the head of the remote branch. The result is that
automated test runs end up testing old revisions of the code, which is
undesirable. To work around this problem, we clone the master branch of
our product repositories rather than a specific branch. This prevents
the Jenkins Git plugin from finding multiple candidate revisions when
determining the revision to build.