Reduce time spent waiting for EC2 datasource
Now that we're using the cloud-init EC2 datasource for our internal VM
artifacts, this adds roughly 100 seconds of additional latency to our
boot time (i.e. the time until ssh is available) for VMs not running in
EC2. This is because the default timeout is 50 seconds, and cloud-init
will attempt to contact 2 endpoints, waiting 50 seconds for each.

This change reduces this timeout to 10 seconds, reducing the total
additional boot time latency to roughly 20 seconds (when not running in
EC2). Further, the max_wait time was reduced to 24 seconds, to maintain
the same ratio of timeout to max_wait time as before (the default
max_wait is 120).