Add the delphix user to the root group
This commit adds the delphix user to the root group as a matter of
convenience. The delphix management service runs as root with root:root
owned log files that are group readable. This minimally allows the
delphix user to be able to read those log files without having to use
sudo.

Closes #11