LX-1100 appliance-build portion of LX-788
When testing OOM conditions in the stack, it was discovered that the
script which creates the hprof dump could fail to be run because the JVM
uses fork(2) rather than vfork(2). In order to allow the dump creation
to proceed, the default settings for vm.overcommit_memory needed to
change from the default.

Additionally, this change enables application core dumps for all users
and creates a common directory in which all cores are written. Cores
will use the same filename pattern as on illumos.

Tested that an hprof dump is successfully created when the stack
encounters an OOM error using the fixed test_out_of_memory blackbox
test.

Created a live build, and tested that cores are created for root and
user processes in the correct location.