Add initial support for building metapackages
This change adds some initial support for building a new metapackage
named "delphix-foundation". The intention is for this package to depend
on other packages, such that this list of dependencies builds the
"foundation" of the appliance. The current list of dependencies is
defined in the "metapackages/seeds/bionic/foundation" file.

This new package isn't yet consumed by anything, but the plan is to
migrate the live-build base configuration over to using this package;
i.e. the files in "live-build/base/config/package-lists" would be
replaced with a single file that only listed "delphix-foundation".

The new package can be easily generated via "make metapackages":

    $ ./scripts/docker-run.sh make metapackages

    $ ls -l *.deb
    -rw-r--r-- 1 root root 1.3K Jun 15 22:07 delphix-foundation_1.0.0_amd64.deb

    $ dpkg-deb -I delphix-foundation_1.0.0_amd64.deb
     new Debian package, version 2.0.
     size 1252 bytes: control archive=628 bytes.
         531 bytes,    10 lines      control
          80 bytes,     1 lines      md5sums
     Package: delphix-foundation
     Source: delphix-meta
     Version: 1.0.0
     Architecture: amd64
     Maintainer: Delphix Engineering <eng@delphix.com>
     Installed-Size: 7
     Depends: atop, auditd, bpfcc-tools, cloud-init, debsums, dnsutils, dstat, emacs, gdb, glances, grub-pc, iftop, inotify-tools, linux-generic, linux-tools-generic, net-tools, open-vm-tools, openssh-server, procinfo, sg3-utils, strace, tshark, ubuntu-minimal, vim, zfs-initramfs, zfsutils-linux
     Section: metapackages
     Priority: optional
     Description: Foundation for Delphix Appliance

In addition to the ".deb" file, these other files are also generated:

    $ ls -l delphix-*
    -rw-r--r-- 1 root root 1.3K Jun 15 22:07 delphix-foundation_1.0.0_amd64.deb
    -rw-r--r-- 1 root root  575 Jun 15 22:07 delphix-meta_1.0.0.dsc
    -rw-r--r-- 1 root root 1.9K Jun 15 22:07 delphix-meta_1.0.0.tar.xz
    -rw-r--r-- 1 root root 5.6K Jun 15 22:07 delphix-meta_1.0.0_amd64.buildinfo
    -rw-r--r-- 1 root root 1.6K Jun 15 22:07 delphix-meta_1.0.0_amd64.changes