WIP: Add inital scripts to build "seed repository"
This change adds some initial logic to build the "seed repository" based
on the dependencies of the "delphix-foundation" package. This is simply
meant to serve as a proof-of-concept, rather than a finished work. After
building the "delphix-foundation" metapackage, that package's recursive
dependencies are determined, and then each dependent package is pulled
down and inserted into a local repository using Aptly.

The intention is for this local Aptly repository to then be used by
live-build, such that our appliance is generated from this local
repository rather than from Ubuntu's repository. Currently though, this
local repository is not consumed by anything.

Further, this concept needs to also be applied to our other delphix
packages (e.g. virtualization, masking), so that we populate this seed
repository with all of the recursive dependencies of those packages.
Additionally, we need to determine how to populate this seed repository
with the various packages currently installed via the appliance-build
Ansible roles (e.g. compilers, tools, etc.).

Some initial testing/examples are below:

    $ ./scripts/docker-run.sh make seed-repository
    ...

    $ ls -l artifacts/
    total 235M
    -rw-r--r-- 1 root root 235M Jun 18 21:08 seed-repository.tar.gz

    $ tar -tf artifacts/seed-repository.tar.gz | egrep '.deb$' | head
    ./pool/5e/e2/ac50a7e374a8312fb68c64da27da_libxrandr2_2%3a1.5.1-1_amd64.deb
    ./pool/5e/cd/8162a82d8c601d8951be2dcd1934_python3-jwt_1.5.3+ds1-1_all.deb
    ./pool/d4/51/eb00b2d0e2940e26028ea9f2fb1b_adduser_3.116ubuntu1_all.deb
    ./pool/ff/c4/8ddd362288e45525f12a9a84eae5_libfreetype6_2.8.1-2ubuntu2_amd64.deb
    ./pool/ff/16/4729742f8e8cfc1024d6b62d9d7d_libpython3.6-stdlib_3.6.5-3_amd64.deb
    ./pool/ff/96/a9900df192e718262fc12ccb0822_python3-markupsafe_1.0-1build1_amd64.deb
    ./pool/8c/06/8498210438759bc0daafbb1cf74d_libgif7_5.1.4-2_amd64.deb
    ./pool/88/02/c7cb7bfa0fda6d899f1cb988dba9_libthai0_0.1.27-2_amd64.deb
    ./pool/88/15/44ee71d85e89ebf224e45acc9493_libxau6_1%3a1.0.8-1_amd64.deb
    ./pool/2e/6d/afb986ee8ebbc4e1c344fab090a4_zlib1g_1%3a1.2.11.dfsg-0ubuntu2_amd64.deb

    $ tar -tf artifacts/seed-repository.tar.gz | egrep '.deb$' | wc -l
    383