Use "ancillary" repository when running live-build
This patch modifies the method in which we download and install the
"first-party" Delphix package into the appliance.

Prior to this change, these packages would be downloaded and installed
via an Ansible role that was executed by the binary hook stage of
live-build. This meant, the initial base configuration of live-build
(e.g. the "minimal.list.chroot" file) could not depend on any Delphix
packages.

With this change, we now download the Delphix packages prior to running
live-build, and build an Aptly/APT package repository containging these
packages. We then configure live-build to point to this new repository
being served over HTTP on the localhost, such that live-build will use
it to pull in any Delphix packages it needs.

As a result of this change, we can now depend on the Delphix ZFS
packages in the base configuration, since the base configuration can
easily pull from this new APT repository containing the Delphix packages.

Co-authored-by: Pavel Zakharov <pavel.zakharov@delphix.com>