Adapt "upgrade-repository" generation logic
Now that we've moved to using the "ancillary" repository to install the
Delphix packages, we also need to update the "upgrade-repository"
generation logic to account for this. This patch makes these changes,
removing the reliance on the old "pkgrepo" directory that was previously
used to store the Delphix packages.

Also, the "delphix-entire" package that's dynamically generated is
installed into the binary chroot directory, so that the resultant
appliance VM artifacts will produce a system with this package
installed. This will likely be necessary based on our current appliance
upgrade architecture plans.

Lastly, we now remove the /etc/apt/sources.* files from the appliance.
When running production systems, we don't want to rely on any external
APT repositories for upgrade or installation of packages, so we remove
these files to help prevent any mistakes and/or accidental use.

While there, some miscellaneous changes were made to the upgrade
repository generation script and package template file, to hopefully
make things more simple and maintainable.

Co-authored-by: Pavel Zakharov <pavel.zakharov@delphix.com>