WIP: Install new "delphix-foundation" package
This change modifies the base live-build configuration, to install a new
"delphix-foundation" package which will depend on all of the packages
that constitute the "foundation" of the Delphix appliance. As a result,
the live-build configuration only needs to install this single package,
rather than listing all of the packages that constitute the "base".