Install snmp package containing client utilities by default.
This commit installs the snmp package to our base appliance image.
The SNMP client utilities included in this packge (e.g. snmpwalk and
snmpget) are useful for the field/support to debug the snmpd daemon.

Additionally, blackbox needs these utilities for test automation, so
installing this package kills two birds with one stone.