Restore default value for ZFS parameter: zfs_vdev_async_write_min_active
The default value for "zfs_vdev_async_write_min_active" is 1 in illumos,
but it's currently 2 in the Linux port. We don't have a specific reason
for going with a value of 1 vs. 2 for DelphixOS, so rather than
customize this tunable to match illumos, we should just use the default
value used in Linux.