Restore default value for ZFS parameter: zfs_vdev_aggregation_limit
The default value for "zfs_vdev_aggregation_limit" is 128K in illumos,
but it's currently 1MB in the Linux port. We don't have a specific
reason for going with a value of 128K vs. 1MB for DelphixOS, so rather
than customize this tunable to match illumos, we should just use the
default value used on Linux.