Temporarily install build dependencies in Virtualization Engine product variants.
As the comment states, we would like to use the "internal-qa" variant
for pre-push testing; however, in order to do that, we would currently
need to revamp our entire pre-push testing workflow. As a workaround, we
temporarily include the build dependencies in the "internal-qa" variant
to avoid having to revamp the pre-push testing workflow. Once the
pre-push testing workflow is redone, we can move the build dependencies
back into the "internal-dev" variant.