Fix LDAP for the "internal-dev" variant
In order for LDAP to work properly, we need a specific ordering of the
values contained in the "passwd" line of the "/etc/nsswitch.conf" file.

Currently, that line looks like this:

    passwd:         compat systemd delphix ldap

As a result, logging in using one's LDAP user does not work. To fix
this, we need this line to look like the following:

    passwd:         compat systemd ldap delphix

Essentially, we need the "delphix" value to always be last.

Due to the way the virtualization package configures this line, we need
to ensure the virtualization package is installed after the LDAP
specific configuration has already been applied. This change acheives
this goal by changing the order in which the appliance-build ansible
roles are applied; i.e. now the LDAP specific role is applied first, and
then the virtualization specific roles are applied later.