Don't use "no_log" with the "user" module
We currently use the "no_log" field when using the "user" module in
Ansible to avoid leaking the contents of the user password. This is
unnecessary since the contents of the password will automatically be
hidden by Ansible, even when running with Ansible's verbose options.

For example, if I remove "no_log" and cause the "user" module to fail,
I see the following output:

    TASK [appliance-build.minimal-common : user]
    ...
    fatal: [binary]: FAILED! => {
        "changed": false,
        "invocation": {
            "module_args": {
    ...
                "password": "VALUE_SPECIFIED_IN_NO_LOG_PARAMETER",
    ...
    }

By no longer using "no_log", we can get more details when the "user"
module fails, which helps us diagnose problems on first failure.