Add passwordless sudo for LDAP users on internal-dev appliances
This works on Illumos appliances, so developers will expect it to work
it less likely that users will overlook bugs where we accidentally rely
on Illumos by restricting passwordless sudo to LDAP users, which makes
on Linux as well. Here we do something slightly better than what we do
on these extra privileges which don't exist on the version we ship.