LX-1358 nfs-mountd --manage-gids bug prevents sybase user from provisioning Oracle VDBs
Disable the --mange-gids flag for nfs-mountd. By default, the service provides
the --manage-gids flag so NFS will be aware of group membership for users in
more than 16 groups. However, this flag has a bug, preventing users from
accessing files owned by a group they are a member of, unless it's their
primary group.