Use "delphix-platform" package and configuration
This change modifies the base live-build configuration, such that we
only list and install a new "delphix-platform" package. This new package
will depend on all the packages that are currently listed in the base
configuration's "minimal" and "tools" package lists. This way, when we
install this new package, we'll still get the same set of packages
installed.

Further, this change creates a new machanism for configuration to be
applied at appliance-build time, but have that configuration supplied by
external packages that are installed in the appliance's root filesystem.
This new mechanism is used so that much of the configuration that was
previously delivered by the "appliance-build.minimal-common" Ansible
role, is now delivered by the new "delphix-platform" package. This way,
this configuration can be applied at appliance-build time, but can also
be properly modified and re-applied on upgrade.