Fix assignment of the AWS_S3_URI_PLATFORM variable
In the "build-ancillary-repository.sh" script, we incorrectly use the
virtualization package's "latest" S3 prefix when assigning a value to
the AWS_S3_URI_PLATFORM variable.

If either the AWS_S3_URI_PLATFORM or AWS_S3_PREFIX_PLATFORM variables
are set, this won't cause a problem, but for "manual" builds that
"latest" prefix will be used (i.e. when neither of those environment
variables are set). As a result, we'll fail to download the
"delphix-platform" package (and add it to the ancillary repository),
which will result in the build failing with the error:

    E: Unable to locate package delphix-platform

The fix is to simply use the proper "latest" S3 prefix for this package.