Add "not-in-place" option on upgrade container creation
This change modifies the script that's used to create the container used
for upgrade verification, adding a new "not-in-place" mode. Thus, now
there's two modes: "in-place" and "not-in-place".

When creating the container "in-place", the container will be created
from a snapshot of the current root filesystem. This mode is the same as
what was previously supported. The result of this, will be a new
container with all of the same software installed as the current root
filesystem.

When creating the container "not-in-place", the container will be
created from scratch by using debootstrap. The result of this, will be a
new container will only a very minimal set of software installed, such
that the container can be started and the "delphix-entire" package can
be installed.

While the "not-in-place" mode is not currently used on upgrade (we still
only use the "in-place" container creation support), the intention is to
build on this, and eventually allow upgrading to a new root filesystem
that is not a clone of the current root filesystem.