Add new "run" option to "upgrade-container" script
This change adds a new "run" option to the "upgrade-container" script
which is a wrapper around the "systemd-run" command. This is meant to
make it slighly easier to run commands in a given upgrade container
during development and/or testing.

This also provides a thin layer of encapsulation between users/scripts
that need to run commands in the container, and the details of actually
doing that (e.g. we could potentially modify the implementation of this
"run" option without modifying the consumers of it).