Infer version from path of script being executed
This changes the interface of the upgrade scripts such that the version
being used for the upgrade or verification does not have to be specified
as an option/parameter to the script. Rather, we now infer the version
from the path of the script being executed, since all of these scripts
will reside in a version specific subdirectory of "/var/dlpx-update".

This eliminates the inconsistency between the scripts, where some would
require the version to be specified using "-v", and others would require
the version to be passed as a positional parameter.