Configure delphix user's password during build
Soon the delphix-platform will remove it's current logic that handles
configuration of the delphix user's password. Thus, we (now) need to
handle configuration of the password in this repository, as part of the
initial appliance build.

Additionally, we're removing the ability to change the appliance's
username, since this flexibility is unnecessary, and only serves to make
things more difficult on upgrade and/or migration.