Always use "delphix.XXXXXXX" as the rootfs name
With this change, the root filesystem(s) of the appliance will take on a
different naming scheme. Currently, the name will be one of: "ubuntu",
"in-place.XXXXXXX", or "not-in-place.XXXXXXX", depending on how the
filesystem was generated (e.g. on initial build, or upgrade). With this
change, now the name will always be "delphix.XXXXXXX" regardless of if
the filesystem was generated on initial appliance build, or on upgrade.