Reload and restart delphix-platform during upgrade
This updates the upgrade logic to account for some recent changes to the
delphix-platform service. Now, we need to reload the service to ensure
the service will do useful work the next time it runs. Additionally, we
need to restart the service rather than start the service; otherwise, if
the service is already running (it should be), starting the service is
basically a noop.