Remove "ansible-done" file during upgrade
This change updates the upgrade logic such that we properly remove the
"delphix-platform" service's "ansible-done" file prior to restarting the
service. Otherwise, restarting the service may not do anything useful,
since the service will only run the new ansible logic if this file
doesn't exist; i.e. if the file exists, the service is basically a noop.