Fix recent regression in "execute" script
The following commit broke the "execute" script:

    commit e3ea5c8dc820ce276553c16ba3ba506540495337
    Author: Prakash Surya <prakash.surya@delphix.com>
    Date:   Thu Dec 13 10:15:40 2018 -0800

        Enforce version.info requirements in upgrade scripts

As an example, running "upgrade" resulted in this error:

    E: Version '' for 'delphix-entire' was not found
    execute: upgrade failed; from '' to ''
    upgrade: '/var/dlpx-update/2018.12.13.7/execute' failed in 'delphix.IPS0gAD'

The problem is we use the "INSTALLED_VERSION" and "VERSION" variables in
the "execute" script, but that commit removed the code that was
responsible for setting those variables. This change resolves the issue
by adding code to set these variables prior to them being used.