Use "/etc/fstab" to mount non-rootfs datasets
We now have non-rootfs datasets that are distinct from the root
filesystem dataset, but are tied to a specific root filesystem. Thus,
when we boot a specific root filesystem, we need to ensure we always
mount the correct non-rootfs datasets that are tied to that given root
filesystem.

This change attempts to accomplish this, by relying on the "/etc/fstab"
file contained in the root filesystem, to mount these non-rootfs
datasets. This way, the file can be unique to each root filesystem,
pointing to the specific non-rootfs datasets that're tied to that root
filesystem, and result in the correct non-rootfs datasets automatically
being mounted at boot time.