Use "/var/log/delphix-upgrade" for upgrade logs
This change modifies the upgrade logic to use a new directory for
storing upgrade log files. This directory is contained on the host's
root filesystem, but is bind mounted into the container. This way,
software running in the upgrade container can write to it, and the files
can be accessed by software running on the host.