Revert "Use zfs-mount.service to mount datasets in container"
This reverts commit 20573a5d4ca5abe537c5f232b5ecf920a58286a0.

The change being reverted breaks not-in-place upgrade. This is due to
how we "debootstrap + start" the upgrade container, prior to installing
the delphix-entire package (and all of it's dependencies).

As a result, when the container is first started, the "zfs-mount"
service will not exist, and thus the "/var/dlpx-update" directory is not
automatically mounted. Then, because this directory isn't mounted, the
upgrade image required to execute the upgrade and install delphix-entire
is not accessible from the container.

The solution is to simply revert the breakage, and continue to rely on
systemd-nspawn to mount this container, since this works even when
delphix-entire is not installed in the container.