Use a dedicated dataset for GRUB configuration
Use a dedicated/seperate dataset for grub, so that we can maintain it
differently than how we maintain and version the root filesystems. This
allows us to have a single configuration for the bootloader, spanning
all of the root filesystems that may be present on the system; rather
than trying to have a configuration on each root filesystem, and trying
to keep them all consistent with each other.