Add script to choose which rootfs to boot next
This changes create a new upgrade script, rootfs-container, which can be
used to configure which "rootfs container" to be used as the appliance's
boot filesystem; i.e. the filesystem that will be used as the rootfs the
next time the system boots.

This logic had previously been contained in the "upgrade-container"
script's "convert_to_bootfs" function, so this change simply moves this
logic into a new script so it can be invoked seperately.

As an example, take the following system:

    $ zfs list /
    NAME                              USED  AVAIL     REFER  MOUNTPOINT
    rpool/ROOT/delphix.a0Jj3pP/root  6.77G  32.2G     6.77G  /

    $ sudo zfs list -r rpool/ROOT -d 1
    NAME                         USED  AVAIL     REFER  MOUNTPOINT
    rpool/ROOT                  31.6G  32.2G       65K  none
    rpool/ROOT/delphix.a0Jj3pP  6.78G  32.2G       64K  none
    rpool/ROOT/delphix.dREvZi4  18.0G  32.2G       64K  none
    rpool/ROOT/delphix.g1RRkAS  6.84G  32.2G       65K  none

We can use this new script to switch the boot filesystem:

    $ sudo /var/dlpx-update/latest/rootfs-container set-bootfs delphix.dREvZi4
    Installing for i386-pc platform.
    Installation finished. No error reported.
    Generating grub configuration file ...
    Found linux image: /boot/vmlinuz-4.15.0-43-generic
    Found initrd image: /boot/initrd.img-4.15.0-43-generic
    done

    $ sudo reboot

    $ zfs list /
    NAME                              USED  AVAIL     REFER  MOUNTPOINT
    rpool/ROOT/delphix.dREvZi4/root  8.21G  31.9G     7.61G  /