Add "rollback" command to "upgrade" script
This change builds on the previously added "set-bootfs" command of the
"rootfs-container" script, now providing an easy way to "rollback" after
booting to an upgraded rootfs container. This allows a user to easily
revert an upgrade, if the upgrade is deemed unsuccessful.

For example, one can run the following to perform an upgrade:

    $ sudo /var/dlpx-update/latest/upgrade -v not-in-place
    $ sudo reboot

Now, after the reboot, if it's determined that the upgrade was not
successful, one can run the following to rollback that prior upgrade:

    $ sudo /var/dlpx-update/latest/upgrade rollback
    $ sudo reboot

At this point, the system will boot to the original root filesystem that
was used prior ot the upgrade.