Support "rollback" for in-place upgrades
This change enables rollback for in-place upgrades. It does this by
creating a new type of upgrade container; a "rollback" upgrade
container. This container type can only be created after an in-place
upgrade has been performed, as it relies on the snapshots that are
created by the "execute" script during an "in-place" upgrade.

During rollback after an in-place upgrade, a "rollback" upgrade
container will be created, and then this container is immediately
converted into the rootfs/bootfs container. Thus, after the rollback
operation and a subsequent reboot, that rollback container will be the
mounted rootfs.