Enable use of main and secondary APT mirrors
This change better enables us to use internal APT package mirrors in the
build, and to dynamically specify the URLs of such mirrors.

Specifically, this allows us to specify a "main" mirror (which is
enabled directly with "lb config"), and a "secondary" mirror (which is
enabled via the "config/archives" directory). Further, the URLs for both
the "main" and "secondary" mirrors can be specified via the two new
environment variables:

 * DELPHIX_PACKAGE_MIRROR_MAIN
 * DELPHIX_PACKAGE_MIRROR_SECONDARY

Closes #319