Require DELPHIX_SIGNATURE_VERSIONS to be passed in
This changes which keys are used when signing the upgrade image.

Previously we used a hardcoded value of 5.3, and thus we always signed
the upgrade image with the 5.3 key. Now, we require the user to specify
a space seperated list of versions, and we sign the upgrade image with
the key corresponding to each version in the list.

This is intended to work inconjunction with some Jenkins changes, where
Jenkins will build the correct list of signature versions, based the
specific version of the upgrade image being built.