Upgrade image file is better left uncompressed
The file we use for the upgrade image is better left uncompressed
because most of its contents is already compressed. Thus, by having the
file compressed as well, we're mostly wasting resources attempting to
compress (and later decompress) data that is already compressed.

If we leave the file uncompressed, it's actually slightly smaller in
size, than the compressed version of the same data:

    $ du -k internal-dev.upgrade.tar*
    5615228 internal-dev.upgrade.tar
    5616954 internal-dev.upgrade.tar.gz

    $ file internal-dev.upgrade.tar*
    internal-dev.upgrade.tar:    POSIX tar archive (GNU)
    internal-dev.upgrade.tar.gz: gzip compressed data, last modified: Tue Aug 27 12:54:30 2019, from Unix

    $ tar --list -f internal-dev.upgrade.tar
    SHA256SUMS.sig.5.3
    SHA256SUMS
    version.info
    prepare
    payload.tar.gz

    $ tar --list -f internal-dev.upgrade.tar.gz
    SHA256SUMS.sig.5.3
    SHA256SUMS
    version.info
    prepare
    payload.tar.gz