Rely on caller to provide upgrade minimum versions
This change modifies how we populate the "version.info" file of an
upgrade image. Rather than having the "minimum version" and "minimum
reboot optinal version" fields be stored directly in the template file,
we now rely on the caller to pass in those files. Generally, this will
be passed in by our Jenkins automation.