Embed version information as root dataset property
We want a way to easily gather appliance version information about the
root filesystems contained on an appliance, and ideally do so without
having to rely on any non-platform component (e.g. not rely on MDS
and/or the virtualization package).

Thus, this change embeds the appliance version into each root filesystem
container using dataset properties. We maintain a property that contains
the version that was initially installed on a given root filesystem
container (this property will not change), and another property that
contains the currently installed version (this may change via upgrades).

This updates the following codepaths to maintain these properties:

 * Initial VM image creation
 * Migration from illumos to Linux
 * Upgrade from Linux to Linux
 * Rollback from Linux to Linux