remove crash-python role from appliance-build
Now that we can run crash-python outside the git repo
from any path and we are halting future development on
it, its role on the appliance build is no longer needed
for developer VMs.

Note that deleting that role uncovered a different bug
where the minimal-common role in internal-dev is needed
by the ldap role and it would always run before it
because it was listed as a dependency under the crash
python role. Now that the crash-python role is removed
we re-add an old version of that configuration making
the minimal-common role run before the ldap role.