Log upgrade scripts' stdout and stderr via syslog
This change builds on the prior commit 9ba7066. Now, in addition to
logging the upgrade scripts' execution trace to syslog, we allow scripts
to log the stdout and stderr of their commands to syslog as well.

This new functionality is leveraged in the "execute" script, such that
we keep a better log of what "apt-get" is doing when it's upgrading
packages on the system.