Add run-pty subcommand to upgrade-container script
This adds a new "run-pty" subcommand to the "ugprade-container" script
which is intended to be used when running interactive commands in a
container; e.g. running a shell. Currently, the "run" subcommand can be
used, but since that does not use the "--pty" option with "systemd-run",
the terminal may not work correctly (e.g. keyboard input, pager output,
etc). Now, when folks want to run interactive commands within a
container, they should use "run-pty" rather than "run", to avoid these
potential issues.