Add S3 inputs to "artifacts" directory
This change helps us preserve the inputs to the build that come from S3.
Previously, after the files from S3 were downloaded and processed, they
would be deleted. Now, we preserve the files within the "artifacts"
file, such that they can be preserved by any automation running the
build, just like we've been doing with the outputs of the build.