Incorrect handling for certain boot devices names (#442)
The "get_bootloader_devices" function attempts to list all devices
attached to "rpool" that are valid devices for storing the bootloader.
Since the devices used for "rpool" are partitions, it has to convert the
partion name (as returned by "zpool list") to the underlying device
name.

Unfortunately, the logic to do that mapping did not properly support
device partiations with "p" in the name; e.g. "nvme0n1p1" should map to
"nvme0n1", and "sda1" to "sda". This change updates this logic to better
support partitions names, regardless of it containing a "p" in the
partiation name, or not.