"rootfs-container delete" must process "log" dataset (#448)
When "/var/log" was moved to a seperate dataset in commit 76a37744, we
failed to properly adapt the "rootfs-container delete" script to account
for this change. Since then, the "rootfs-container delete" script has
likely been broken; currently, when I attempt to use this script, it
fails with the following error:

    $ sudo /var/dlpx-update/latest/rootfs-container delete delphix.xfXCbRc
    cannot destroy 'rpool/ROOT/delphix.xfXCbRc': filesystem has dependent clones
    use '-R' to destroy the following datasets:
    rpool/ROOT/delphix.iluoSLh/log
    rootfs-container: 'zfs destroy -r rpool/ROOT/delphix.xfXCbRc' failed

The problem is, we're not properly handling the "log" dataset clones.
The fix is simple, we need to add the "log" dataset to the list of
datasets that we account for (i.e. the datasets that may need to be
"zfs promote"-ed) when calling "rootfs-container delete".