Backports of 4 commits to add 'upgrade finalize' to 6.0/stage (#457)
* [Backport to 6.0.2.0] Add new "finalize" subcommand to upgrade script (#413)

This change adds the "finalize" subcommand to the upgrade script, which
is intended to be used to clean up any filesystems, snapshots, etc. that
may have been generated during the upgrade process (e.g. for rollback).

This is intended to be used to mark the finalization or completion of
the upgrade, after which rollback will no longer be possible; e.g.

    $ sudo /var/dlpx-update/latest/upgrade deferred
    $ sudo /var/dlpx-update/latest/upgrade finalize

* [Backport to 6.0.2.0] "rootfs-container delete" must process "log" dataset (#448)

When "/var/log" was moved to a seperate dataset in commit 76a37744, we
failed to properly adapt the "rootfs-container delete" script to account
for this change. Since then, the "rootfs-container delete" script has
likely been broken; currently, when I attempt to use this script, it
fails with the following error:

    $ sudo /var/dlpx-update/latest/rootfs-container delete delphix.xfXCbRc
    cannot destroy 'rpool/ROOT/delphix.xfXCbRc': filesystem has dependent clones
    use '-R' to destroy the following datasets:
    rpool/ROOT/delphix.iluoSLh/log
    rootfs-container: 'zfs destroy -r rpool/ROOT/delphix.xfXCbRc' failed

The problem is, we're not properly handling the "log" dataset clones.
The fix is simple, we need to add the "log" dataset to the list of
datasets that we account for (i.e. the datasets that may need to be
"zfs promote"-ed) when calling "rootfs-container delete".

* [Backport to 6.0.2.0] Remove lingering snapshots after rollback event (#449)

After performing a rollback of a failed not-in-place upgrade, if we remove
the container associated with that failure (i.e. the "new" container), we
will not remove the "container-delphix.XXXXXXX" snapshots used to generate
that new container, and these snapshots will persist indefinitely. This
change adds the necessary logic to the "rootfs-container delete" script to
remove these snapshots, after removing the container itself.

For example, without this change:

    $ download-latest-image internal-dev
    $ sudo unpack-image -x internal-dev.upgrade.tar
    $ sudo /var/dlpx-update/latest/upgrade -v full
    $ sudo /var/dlpx-update/latest/upgrade rollback
    $ sudo /var/dlpx-update/latest/rootfs-container delete delphix.Bs4VTPo
    $ sudo zfs list -t all -r rpool/ROOT/delphix.jDHoOtU
    NAME                                                        USED  AVAIL     REFER  MOUNTPOINT
    rpool/ROOT/delphix.jDHoOtU                                 17.7G  44.3G       64K  none
    rpool/ROOT/delphix.jDHoOtU@container-delphix.Bs4VTPo          0B      -       64K  -
    rpool/ROOT/delphix.jDHoOtU/data                            34.1M  44.3G     33.1M  legacy
    rpool/ROOT/delphix.jDHoOtU/data@container-delphix.Bs4VTPo  1006K      -     30.5M  -
    rpool/ROOT/delphix.jDHoOtU/home                            11.8G  44.3G     11.8G  legacy
    rpool/ROOT/delphix.jDHoOtU/home@container-delphix.Bs4VTPo   151K      -     11.8G  -
    rpool/ROOT/delphix.jDHoOtU/log                             28.7M  44.3G     27.0M  legacy
    rpool/ROOT/delphix.jDHoOtU/log@container-delphix.Bs4VTPo   1.61M      -     3.95M  -
    rpool/ROOT/delphix.jDHoOtU/root                            5.86G  44.3G     5.86G  /
    rpool/ROOT/delphix.jDHoOtU/root@container-delphix.Bs4VTPo  2.02M      -     5.86G  -

And now, with this change:

    $ download-latest-image internal-dev
    $ sudo unpack-image -x internal-dev.upgrade.tar
    $ sudo /var/dlpx-update/latest/upgrade -v full
    $ sudo /var/dlpx-update/latest/upgrade rollback
    $ sudo /var/dlpx-update/latest/rootfs-container delete delphix.lJvMWgM
    $ sudo zfs list -t all -r rpool/ROOT/delphix.jDHoOtU
    NAME                              USED  AVAIL     REFER  MOUNTPOINT
    rpool/ROOT/delphix.jDHoOtU       17.7G  44.3G       64K  none
    rpool/ROOT/delphix.jDHoOtU/data  34.5M  44.3G     34.5M  legacy
    rpool/ROOT/delphix.jDHoOtU/home  11.8G  44.3G     11.8G  legacy
    rpool/ROOT/delphix.jDHoOtU/log   27.1M  44.3G     27.1M  legacy
    rpool/ROOT/delphix.jDHoOtU/root  5.86G  44.3G     5.86G  /

Thus, with this change, the "container-delphix.XXXXXXX" snapshots are
properly removed after calling "rootfs-container delete".

* DLPX-69176 [Backport of DLPX-68157 to 6.0.2.0] Update 'finalize' logic to clean up snapshots older than current and previous

Co-authored-by: Prakash Surya <prakash.surya@delphix.com>